Fixed legacy unit tests
The Compiler unit test was fixed by adding join information to MySQL Table
objects created in the Select Statement class. This solved all but one scenario,
the testGetChildren test for the Compiler.

Looking at the comments for the test scenario, and understanding the intent
of the data structure, I've determined that the test itself was incorrect -
if a Family can only be constructed from a Parent and a Child, then any
join to a Child must be an INNER JOIN; the request itself LEFT JOINs against
the Family table, and so the intent of getting a list of Child for a given
Parent should be acheived to include an empty list but a populated Parent.

The Php/Output and Legacy/Parser tests did not appear to be broken.